+++ AdincoPetrovendorDB/AdincoDB/dbo/Stored Procedures/Procs3/SP_SE_A7_MPY.sql
@@ -10,7 +10,7 @@
     DROP PROCEDURE SP_SE_A7_MPY;
 GO
 
-CREATE PROCEDURE [dbo].[SP_SE_A7_MPY] 
+CREATE PROCEDURE [dbo].[SP_SE_A7_MPY]
     @IdContrato    INT,
     @IdUsuario     INT,
     @IdPresupuesto INT,
@@ -30,7 +30,7 @@ AS
                 UUID                  VARCHAR(500),
                 RFC                   VARCHAR(50),
                 EncontradoPetrovendor INT,
-				 SubTotal              money,
+				SubTotal              MONEY,
                 IdSubcontratista      INT,
                 IdContrato            INT,
                 IdMoneda              INT,
@@ -39,7 +39,7 @@ AS
         CREATE TABLE #DATOS
             (
                 Descripcion VARCHAR(50),
-                SubTotal    FLOAT,
+                SubTotal    MONEY,
                 IdFactura   INT
             )
 		DECLARE
